+++ app/forms/balance/new_form.rb
@@ -1,7 +1,0 @@
-class Balance::NewForm < Balance::Form
-  property :currency_id
-
-  validation do
-    required(:currency_id).filled
-  end
-end